Store and round durations to the nearest second

+++ app/services/schedule_builder.rb
@@ -41,7 +41,7 @@ class ScheduleBuilder
   private_class_method def self.set_minutes_per_week(schedule:, value:)
     return schedule unless value
 
-    value = value.to_i
+    value = value.to_f
     max_minutes_per_day = 480.0 # 8 hours.
     number_of_days = 7
